fastapi/requirements.txt
fastapi
uvicorn
pydantic
pymongo
email-validator 
dnspython 
requests
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
jinja2